requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
distlib==0.4.0
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-timezone-field==7.1
django_celery_results==2.6.0
docopt==0.6.2
et_xmlfile==2.0.0
filelock==3.20.3
fonttools==4.60.0
idna==3.10
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.6
mysql-connector-python==9.5.0
mysqlclient==2.2.8
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pipreqs==0.4.13
platformdirs==4.5.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.36.1
wcwidth==0.2.14
yarg==0.1.10
gunicorn
whitenoise
dj-database-url
